Updated models, repositories, config, templates, and resources

+++ src/main/java/mk/ukim/finki/easyfood/repository/CartItemsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CartItemsRepository extends JpaRepository<CartItems, Integer> {
+public interface CartItemsRepository extends JpaRepository<CartItems, Long> {
 }
