+++ src/main/java/mk/ukim/finki/easyfood/repository/CostsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CostsRepository extends JpaRepository<Costs, Integer> {
+public interface CostsRepository extends JpaRepository<Costs, Long> {
 }
