+++ src/main/java/mk/ukim/finki/easyfood/repository/DeliveryManRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryManRepository extends JpaRepository<DeliveryMan, Integer> {
+public interface DeliveryManRepository extends JpaRepository<DeliveryMan, Long> {
 }
